Move samples test cases to subdirectory
